+++ workspace/plugins/inbox/src/media/index.ts
@@ -15,29 +15,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import deepmerge from "deepmerge";
-import { defineConfig, UserConfig } from "vite";
+/**
+ * REF: https://github.com/Lykos94/media-socket
+ * REF: https://www.npmjs.com/package/media-socket
+ */
 
-import viteShareConfig from "./vite.share.config";
-import vitePluginConfig from "./vite.plugin.config";
-import viteAppsConfig from "./vite.apps.config";
-
-// https://vitejs.dev/config/
-export default defineConfig(async env => {
-    // console.log(env);
-    var config: UserConfig;
-
-    switch (env.mode) {
-        case "apps":
-            config = deepmerge.all<UserConfig>([viteShareConfig, viteAppsConfig]);
-            break;
-
-        case "plugin":
-        default:
-            config = deepmerge.all<UserConfig>([viteShareConfig, vitePluginConfig]);
-            break;
-    }
-
-    // console.log(config);
-    return config;
-});
+export interface IMediaStreamConstraints extends MediaStreamConstraints {
+    audio: boolean | MediaTrackConstraints;
+    video: boolean | MediaTrackConstraints;
+}